Python3: use int instead of long
Using int for long int should not cause issues since python2.4.
c.f. https://docs.python.org/2.6/whatsnew/2.4.html#pep-237-unifying-long-integers-and-integers

+++ ryu/services/protocols/bgp/utils/bgp.py
@@ -94,7 +94,7 @@ def from_inet_ptoi(bgp_id):
     four_byte_id = None
     try:
         packed_byte = socket.inet_pton(socket.AF_INET, bgp_id)
-        four_byte_id = long(packed_byte.encode('hex'), 16)
+        four_byte_id = int(packed_byte.encode('hex'), 16)
     except ValueError:
         LOG.debug('Invalid bgp id given for conversion to integer value %s',
                   bgp_id)
